ocdi/requirements.txt
scipy
wandb
matplotlib
numpy
prettytable
torchsummary
torchinfo
seaborn
pandas
scikit-learn
torchvision
